Move randomized test generators to incrementalmerkletree
