+++ bridgetree/src/lib.rs
@@ -1299,8 +1299,8 @@ mod tests {
     use proptest::prelude::*;
 
     use super::*;
-    use crate::testing::{apply_operation, arb_operation, tests};
-    use incrementalmerkletree::testing::{Frontier, Tree};
+    use crate::testing::tests;
+    use incrementalmerkletree::testing::{apply_operation, arb_operation, Frontier, Tree};
 
     impl<H: Hashable + Clone, const DEPTH: u8> Frontier<H> for super::Frontier<H, DEPTH> {
         fn append(&mut self, value: &H) -> bool {
