+++ bridgetree/src/testing.rs
@@ -1,137 +1,5 @@
 mod complete_tree;
 
-use proptest::prelude::*;
-
-use incrementalmerkletree::{
-    testing::{SipHashable, Tree},
-    Hashable, Position,
-};
-
-//
-// Types and utilities for shared example tests.
-//
-
-//
-// Operations
-//
-
-#[derive(Clone, Debug)]
-pub enum Operation<A> {
-    Append(A),
-    CurrentPosition,
-    CurrentLeaf,
-    Mark,
-    MarkedLeaf(Position),
-    MarkedPositions,
-    Unmark(Position),
-    Checkpoint,
-    Rewind,
-    Authpath(Position, usize),
-    GarbageCollect,
-}
-
-use Operation::*;
-
-impl<H: Hashable> Operation<H> {
-    pub fn apply<T: Tree<H>>(&self, tree: &mut T) -> Option<(Position, Vec<H>)> {
-        match self {
-            Append(a) => {
-                assert!(tree.append(a), "append failed");
-                None
-            }
-            CurrentPosition => None,
-            CurrentLeaf => None,
-            Mark => {
-                assert!(tree.mark().is_some(), "mark failed");
-                None
-            }
-            MarkedLeaf(_) => None,
-            MarkedPositions => None,
-            Unmark(p) => {
-                assert!(tree.remove_mark(*p), "remove mark failed");
-                None
-            }
-            Checkpoint => {
-                tree.checkpoint();
-                None
-            }
-            Rewind => {
-                assert!(tree.rewind(), "rewind failed");
-                None
-            }
-            Authpath(p, d) => tree
-                .root(*d)
-                .and_then(|root| tree.witness(*p, &root))
-                .map(|xs| (*p, xs)),
-            GarbageCollect => None,
-        }
-    }
-
-    pub fn apply_all<T: Tree<H>>(ops: &[Operation<H>], tree: &mut T) -> Option<(Position, Vec<H>)> {
-        let mut result = None;
-        for op in ops {
-            result = op.apply(tree);
-        }
-        result
-    }
-}
-
-pub fn arb_operation<G: Strategy + Clone>(
-    item_gen: G,
-    pos_gen: impl Strategy<Value = usize> + Clone,
-) -> impl Strategy<Value = Operation<G::Value>>
-where
-    G::Value: Clone + 'static,
-{
-    prop_oneof![
-        item_gen.prop_map(Operation::Append),
-        Just(Operation::Mark),
-        prop_oneof![
-            Just(Operation::CurrentLeaf),
-            Just(Operation::CurrentPosition),
-            Just(Operation::MarkedPositions),
-        ],
-        Just(Operation::GarbageCollect),
-        pos_gen
-            .clone()
-            .prop_map(|i| Operation::MarkedLeaf(Position::from(i))),
-        pos_gen
-            .clone()
-            .prop_map(|i| Operation::Unmark(Position::from(i))),
-        Just(Operation::Checkpoint),
-        Just(Operation::Rewind),
-        pos_gen
-            .prop_flat_map(|i| (0usize..10)
-                .prop_map(move |depth| Operation::Authpath(Position::from(i), depth))),
-    ]
-}
-
-pub fn apply_operation<H, T: Tree<H>>(tree: &mut T, op: Operation<H>) {
-    match op {
-        Append(value) => {
-            tree.append(&value);
-        }
-        Mark => {
-            tree.mark();
-        }
-        Unmark(position) => {
-            tree.remove_mark(position);
-        }
-        Checkpoint => {
-            tree.checkpoint();
-        }
-        Rewind => {
-            tree.rewind();
-        }
-        CurrentPosition => {}
-        CurrentLeaf => {}
-        Authpath(_, _) => {}
-        MarkedLeaf(_) => {}
-        MarkedPositions => {}
-        GarbageCollect => {}
-    }
-}
-
 #[cfg(test)]
 pub(crate) mod tests {
     use proptest::prelude::*;
@@ -139,15 +7,12 @@ pub(crate) mod tests {
     use std::fmt::Debug;
 
     use crate::BridgeTree;
-    use incrementalmerkletree::{Hashable, Level, Position};
-
-    use super::{
-        arb_operation,
-        complete_tree::{lazy_root, CompleteTree},
-        Operation,
-        Operation::*,
-        SipHashable, Tree,
+    use incrementalmerkletree::{
+        testing::{arb_operation, Operation, Operation::*, SipHashable, Tree},
+        Hashable, Level, Position,
     };
+
+    use super::complete_tree::{lazy_root, CompleteTree};
 
     //
     // Shared example tests
